+++ webui-src/app/people/people.js
@@ -3,18 +3,6 @@ const rs = require('rswebui');
 
 const peopleUtil = require('people/people_util');
 
-const UserAvatar = () => ({
-  view: (v) => {
-    const imageURI = v.attrs.avatar;
-    console.log(imageURI);
-    return imageURI === undefined || imageURI.mData.base64 === ''
-      ? []
-      : m('img.avatar', {
-          src: 'data:image/png;base64,' + imageURI.mData.base64,
-        });
-  },
-});
-
 const contactInfo = () => {
   let details = {};
 
@@ -38,7 +26,7 @@ const contactInfo = () => {
         },
         [
           m('h4', details.mNickname),
-          m(UserAvatar, { avatar: details.mAvatar }),
+          m(peopleUtil.UserAvatar, { avatar: details.mAvatar }),
           m('.details', [
             m('p', 'ID:'),
             m('p', details.mId),
